requirements.txt
Django==5.0.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
whitenoise==6.6.0
gunicorn==22.0.0
cuid2==2.0.0
python-dateutil==2.9.0.post0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
